remove make function, use derivateA lazily
