+++ src/graph/index.d.ts
@@ -75,8 +75,6 @@ export function Command<A extends IOTSParams, R, I1 extends Invalidates>(
 export type Queries = { [k: string]: QueryReturn<any, any> }
 export type Commands = { [k: string]: CommandReturn<any, any> };
 
-export function make(graph: Queries): Queries
-
 export function query(...args: any[]): any
 export function runCommand<R>(graph: Queries, command: CommandReturn<any, R>, params: {}): Promise<R>
 export function invalidate(graph: Queries, Ps: (keyof Queries)[], params: {}): void
