+++ src/graph/index.js
@@ -1,4 +1,3 @@
-export { make } from './make';
 export { query, querySync } from './query';
 export { invalidate } from './invalidate';
 export { Query } from './QueryNode';
