+++ src/graph/util.js
@@ -4,6 +4,31 @@ import assign from 'lodash/assign';
 import findKey from 'lodash/findKey';
 import sortBy from 'lodash/sortBy';
 
+export function findP(graph, fetch) {
+  return findKey(graph, { fetch });
+}
+
+function derivateA(graph, P) {
+  const node = graph[P];
+  return node.A || (() => {
+    switch (node.fetch.type) {
+      case 'composition':
+        // in a composition, the "total A" is the master's A
+        // https://github.com/buildo/avenger#composition
+        return derivateA(graph, findP(graph, node.fetch.master));
+      case 'product':
+        // in a product, "total A" is [A1, ..., An]
+        // https://github.com/buildo/avenger#product
+        // In practice, given the "deep" array, we can easily pick and redistribute
+        // the correct pieces of A when given a flat A values object
+        return node.fetch.fetches.map(f => derivateA(graph, findP(graph, f)));
+      default:
+        // a fetch that is neither a composition or a product, **must** carry A information
+        throw new Error(`missing A for naked '${P}'`);
+    }
+  })();
+}
+
 // given a flat `AA` object, e.g.
 //
 //   { a: 3, foo: 'asd', bar: true }
@@ -34,13 +59,9 @@ export function queriesAndArgs(graph, Ps, A) {
     [P]: graph[P].cachedFetch || graph[P].fetch // non-atoms are not cached, we return the naked fetch
   }), {});
   const args = Ps.reduce((argz, P) => assign(argz, {
-    [P]: pickA(A, graph[P].A)
+    [P]: pickA(A, derivateA(graph, P))
   }), {});
   return { queries, args };
-}
-
-export function findP(graph, fetch) {
-  return findKey(graph, { fetch });
 }
 
 // e.g. if `C` depends on `B` depends on `A`
